# Use centroid algorithm with better numerical stability