## Description

This replaces the existing centroid calculation algorithm with one that has better numerical stability. The numerical stability issues were manifesting themselves when calculating the area of small objects (e.g. a few square meters) when projected in WGS84.

E.g. `POLYGON((151 -33,151.00001 -33,151.00001 -33.00001,151 -33.00001,151 -33))", "POINT(151.000005 -33.000005)` would have its centroid calculated as being 10s of kms away from the true centroid.

I also refactored the GeometryCollection code for Centroid, since it felt a bit weird to have point/linear geometries go through a generic function, but areal geometries being treated separately (the 3 cases are now treated separately).

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) N/A, existing function.

## Related Issue

- N/A

## Benchmark Results

As expected, change:

```
name                                          old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.21µs ±11%    2.22µs ± 7%     ~     (p=0.801 n=13+13)
IntersectsLineStringWithLineString/n=100-4      39.2µs ± 6%    39.5µs ±11%     ~     (p=0.867 n=13+14)
IntersectsLineStringWithLineString/n=1000-4      578µs ±10%     594µs ± 7%   +2.75%  (p=0.024 n=14+14)
IntersectsLineStringWithLineString/n=10000-4    9.42ms ± 7%    9.46ms ± 8%     ~     (p=0.983 n=15+14)
LineStringIsSimpleZigZag/10-4                   2.12µs ± 4%    2.12µs ± 6%     ~     (p=0.906 n=14+15)
LineStringIsSimpleZigZag/100-4                  62.1µs ± 4%    62.3µs ± 5%     ~     (p=0.874 n=14+14)
LineStringIsSimpleZigZag/1000-4                 1.00ms ± 3%    1.01ms ± 7%     ~     (p=0.400 n=15+14)
LineStringIsSimpleZigZag/10000-4                13.5ms ± 8%    13.2ms ± 6%     ~     (p=0.104 n=14+14)
PolygonSingleRingValidation/n=10-4              2.45µs ±10%    2.42µs ± 8%     ~     (p=0.471 n=14+15)
PolygonSingleRingValidation/n=100-4             55.1µs ± 3%    55.6µs ± 3%     ~     (p=0.354 n=15+14)
PolygonSingleRingValidation/n=1000-4            1.04ms ± 4%    1.03ms ± 3%     ~     (p=0.352 n=14+14)
PolygonSingleRingValidation/n=10000-4           14.0ms ± 4%    14.1ms ± 5%     ~     (p=0.914 n=14+15)
PolygonMultipleRingsValidation/n=4-4            10.6µs ± 3%    10.9µs ± 6%   +2.61%  (p=0.025 n=14+14)
PolygonMultipleRingsValidation/n=36-4           97.7µs ± 7%    96.6µs ± 3%     ~     (p=0.880 n=15+14)
PolygonMultipleRingsValidation/n=400-4          1.24ms ± 3%    1.23ms ± 3%     ~     (p=0.093 n=14+15)
PolygonMultipleRingsValidation/n=4096-4         14.9ms ± 9%    14.7ms ±10%     ~     (p=0.436 n=15+15)
PolygonZigZagRingsValidation/n=10-4             24.6µs ± 5%    25.0µs ± 4%     ~     (p=0.123 n=14+15)
PolygonZigZagRingsValidation/n=100-4             336µs ± 6%     340µs ± 6%     ~     (p=0.310 n=15+14)
PolygonZigZagRingsValidation/n=1000-4           4.62ms ± 4%    4.58ms ± 3%     ~     (p=0.252 n=15+14)
PolygonZigZagRingsValidation/n=10000-4          62.0ms ± 2%    63.5ms ± 6%   +2.38%  (p=0.008 n=14+15)
MultipolygonValidation/n=1-4                     349ns ± 6%     338ns ± 7%   -2.88%  (p=0.023 n=13+14)
MultipolygonValidation/n=4-4                     903ns ±14%     864ns ± 7%   -4.24%  (p=0.008 n=14+15)
MultipolygonValidation/n=16-4                   5.79µs ±12%    5.66µs ± 8%     ~     (p=0.158 n=14+15)
MultipolygonValidation/n=64-4                   33.9µs ±11%    33.4µs ± 3%     ~     (p=0.667 n=14+14)
MultipolygonValidation/n=256-4                   202µs ± 6%     206µs ±10%     ~     (p=0.412 n=15+15)
MultipolygonValidation/n=1024-4                 1.01ms ± 3%    1.00ms ± 5%     ~     (p=0.546 n=10+14)
MultiPolygonTwoCircles/n=10-4                   7.43µs ± 9%    7.55µs ± 8%     ~     (p=0.300 n=13+15)
MultiPolygonTwoCircles/n=100-4                   104µs ± 7%     104µs ± 6%     ~     (p=0.685 n=14+13)
MultiPolygonTwoCircles/n=1000-4                 1.52ms ± 3%    1.52ms ± 5%     ~     (p=0.982 n=14+14)
MultiPolygonTwoCircles/n=10000-4                22.6ms ±12%    22.2ms ± 7%     ~     (p=0.856 n=15+13)
MultiPolygonMultipleTouchingPoints/n=1-4        6.28µs ±22%    6.13µs ±12%     ~     (p=0.793 n=14+13)
MultiPolygonMultipleTouchingPoints/n=10-4       49.5µs ± 8%    50.5µs ±13%     ~     (p=0.470 n=15+14)
MultiPolygonMultipleTouchingPoints/n=100-4       596µs ± 2%     601µs ± 5%     ~     (p=0.302 n=13+14)
MultiPolygonMultipleTouchingPoints/n=1000-4     7.57ms ± 6%    7.58ms ± 4%     ~     (p=0.801 n=13+13)
MarshalWKB/polygon/n=10-4                        202ns ± 6%     195ns ± 8%   -3.43%  (p=0.026 n=13+14)
MarshalWKB/polygon/n=100-4                       599ns ±13%     568ns ± 9%   -5.12%  (p=0.038 n=14+14)
MarshalWKB/polygon/n=1000-4                     3.83µs ±17%    4.01µs ±38%     ~     (p=0.401 n=14+14)
MarshalWKB/polygon/n=10000-4                    29.5µs ±24%    30.7µs ±23%     ~     (p=0.352 n=14+14)
UnmarshalWKB/polygon/n=10-4                      322ns ± 5%     323ns ±11%     ~     (p=0.553 n=15+14)
UnmarshalWKB/polygon/n=100-4                     710ns ± 8%     729ns ±12%     ~     (p=0.300 n=12+15)
UnmarshalWKB/polygon/n=1000-4                   4.67µs ±22%    4.02µs ± 7%  -13.85%  (p=0.003 n=15+12)
UnmarshalWKB/polygon/n=10000-4                  36.4µs ±14%    35.5µs ±16%     ~     (p=0.413 n=15+13)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             48.4µs ± 5%    48.9µs ±12%     ~     (p=0.899 n=12+14)
Intersection/n=100-4                            93.0µs ± 4%    93.9µs ± 6%     ~     (p=0.477 n=15+14)
Intersection/n=1000-4                            403µs ±10%     413µs ±10%     ~     (p=0.233 n=15+15)
Intersection/n=10000-4                          3.56ms ± 3%    3.59ms ±13%     ~     (p=0.756 n=13+14)
NoOp/n=10-4                                     3.87µs ± 9%    3.89µs ± 8%     ~     (p=0.902 n=15+15)
NoOp/n=100-4                                    12.5µs ± 7%    12.4µs ± 4%     ~     (p=0.683 n=14+15)
NoOp/n=1000-4                                   89.7µs ± 5%    89.1µs ± 6%     ~     (p=0.306 n=14+14)
NoOp/n=10000-4                                  1.00ms ±44%    0.95ms ±18%     ~     (p=0.512 n=15+15)

name                                          old alloc/op   new alloc/op   delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.69kB ± 0%    2.69kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100-4      32.6kB ± 0%    32.6kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000-4      286kB ± 0%     286kB ± 0%     ~     (p=0.088 n=15+15)
IntersectsLineStringWithLineString/n=10000-4    3.77MB ± 0%    3.77MB ± 0%   +0.00%  (p=0.007 n=13+14)
LineStringIsSimpleZigZag/10-4                   1.44kB ± 0%    1.44kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                  21.6kB ± 0%    21.6kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                  230kB ± 0%     230kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000-4                2.30MB ± 0%    2.30MB ± 0%   +0.00%  (p=0.026 n=12+14)
PolygonSingleRingValidation/n=10-4              1.47kB ± 0%    1.47kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4             16.2kB ± 0%    16.2kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4             217kB ± 0%     217kB ± 0%     ~     (p=0.572 n=14+13)
PolygonSingleRingValidation/n=10000-4           2.23MB ± 0%    2.23MB ± 0%     ~     (p=0.606 n=14+15)
PolygonMultipleRingsValidation/n=4-4            5.28kB ± 0%    5.28kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4           43.3kB ± 0%    43.3kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           476kB ± 0%     476kB ± 0%     ~     (p=0.102 n=15+13)
PolygonMultipleRingsValidation/n=4096-4         4.88MB ± 0%    4.88MB ± 0%     ~     (p=0.921 n=14+15)
PolygonZigZagRingsValidation/n=10-4             14.0kB ± 0%    14.0kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100-4             138kB ± 0%     138kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000-4           1.26MB ± 0%    1.26MB ± 0%     ~     (p=0.240 n=14+14)
PolygonZigZagRingsValidation/n=10000-4          15.3MB ± 0%    15.3MB ± 0%     ~     (p=0.356 n=15+15)
MultipolygonValidation/n=1-4                      385B ± 0%      385B ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      676B ± 0%      676B ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                   3.86kB ± 0%    3.86kB ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                   15.4kB ± 0%    15.4kB ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                  63.1kB ± 0%    63.1kB ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  259kB ± 0%     259kB ± 0%     ~     (p=1.000 n=13+13)
MultiPolygonTwoCircles/n=10-4                   6.88kB ± 0%    6.88kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100-4                  59.4kB ± 0%    59.4kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000-4                  508kB ± 0%     508kB ± 0%     ~     (p=0.674 n=15+15)
MultiPolygonTwoCircles/n=10000-4                6.89MB ± 0%    6.89MB ± 0%     ~     (p=0.754 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1-4        4.18kB ± 0%    4.18kB ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10-4       23.8kB ± 0%    23.8kB ± 0%     ~     (p=0.377 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       212kB ± 0%     212kB ± 0%     ~     (p=0.418 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000-4     2.00MB ± 0%    2.00MB ± 0%     ~     (p=0.976 n=15+15)
MarshalWKB/polygon/n=10-4                         232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                      1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                     16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                     164kB ± 0%     164kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       282B ± 0%      282B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                    1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                   16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                   164kB ± 0%     164kB ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             1.19kB ± 0%    1.19kB ± 0%     ~     (all equal)
Intersection/n=100-4                            6.33kB ± 0%    6.33kB ± 0%     ~     (p=1.000 n=15+15)
Intersection/n=1000-4                           55.0kB ± 0%    55.0kB ± 0%     ~     (all equal)
Intersection/n=10000-4                           557kB ± 0%     557kB ± 0%     ~     (p=0.933 n=14+13)
NoOp/n=10-4                                       864B ± 0%      864B ± 0%     ~     (all equal)
NoOp/n=100-4                                    5.68kB ± 0%    5.68kB ± 0%     ~     (all equal)
NoOp/n=1000-4                                   49.5kB ± 0%    49.5kB ± 0%     ~     (all equal)
NoOp/n=10000-4                                   492kB ± 0%     492kB ± 0%   -0.00%  (p=0.034 n=15+15)

name                                          old allocs/op  new allocs/op  delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4         15.0 ± 0%      15.0 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100-4         160 ± 0%       160 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000-4      1.28k ± 0%     1.28k ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10000-4     19.3k ± 0%     19.3k ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10-4                     5.00 ± 0%      5.00 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                    75.0 ± 0%      75.0 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                    797 ± 0%       797 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000-4                 8.00k ± 0%     8.00k ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10-4                6.00 ± 0%      6.00 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4               57.0 ± 0%      57.0 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4               755 ± 0%       755 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10000-4            7.73k ± 0%     7.73k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4-4              25.0 ± 0%      25.0 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4              203 ± 0%       203 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           2.23k ± 0%     2.23k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4096-4          22.8k ± 0%     22.8k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10-4               70.0 ± 0%      70.0 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100-4               636 ± 0%       636 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000-4            5.46k ± 0%     5.46k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10000-4           74.1k ± 0%     74.1k ± 0%   -0.00%  (p=0.020 n=15+12)
MultipolygonValidation/n=1-4                      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                     27.0 ± 0%      27.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                     99.0 ± 0%      99.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                     392 ± 0%       392 ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  1.58k ± 0%     1.58k ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10-4                     46.0 ± 0%      46.0 ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100-4                     326 ± 0%       326 ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000-4                  2.57k ± 0%     2.57k ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10000-4                 38.6k ± 0%     38.6k ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=1-4          51.0 ± 0%      51.0 ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10-4          332 ± 0%       332 ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=100-4       3.07k ± 0%     3.07k ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=1000-4      29.2k ± 0%     29.2k ± 0%     ~     (p=0.173 n=15+13)
MarshalWKB/polygon/n=10-4                         6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                        6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                     7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                    7.00 ± 0%      7.00 ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                               31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=100-4                              31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=1000-4                             31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=10000-4                            31.0 ± 0%      31.0 ± 0%     ~     (all equal)
NoOp/n=10-4                                       21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=100-4                                      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=1000-4                                     21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=10000-4                                    21.0 ± 0%      21.0 ± 0%     ~     (all equal)
```